Working directory hack
Rendering parts of the document in RStudio can fail because RStudio will always run each chunk with the working directory set to the same directory as the Rmd file. In general, we want the working directory to always be the project root directory. This line of code does this, but it is a hack because it only sets it for the current chunk.

Setting knit options doesn't help us here because that's for control of the Knit HTML procedure, not the RStudio preview render mode.